Creating option to rename folders instead of files
Checkbox has been added to the MainWindow so that one can choose to rename
folders instead of files. When checked/unchecked the list of files (or
folders) will be automatically fetched as well. Had to modify the Renamer
adding a boolean to say if we are renaming directories or files and modify
the file filter used (it could not be static anymore).